modify test codes of Query, QueryOperator
